Bugfix: -cs overflow
E.g. -cs 6000 overflowed. It is processed as integer arithmetics 6000 * 1024 * 1024, and only then assigned to size_t; but by then it already overflowed.

+++ src/src_sharpSAT/main.cpp
@@ -225,7 +225,7 @@ int main(int argc, char *argv[])
 				toSTDOUT("wrong parameters"<<endl);
 				return -1;
 			}
-			CSolverConf::maxCacheSize = atoi(argv[i + 1]) * 1024 * 1024;
+			CSolverConf::maxCacheSize = ((size_t) atoi(argv[i + 1])) * 1024 * 1024;
 			//cout <<"maxCacheSize:" <<CSolverConf::maxCacheSize<<"bytes\n";
 		}
 		else
